+++ src/messages/de/admin.php
@@ -367,5 +367,6 @@ return [
     'crud_exportdata_btn_generateexport' => 'Export generieren',
     'crud_exportdata_btn_downloadexport' => 'Export herunterladen',
     'js_pagination_page' => 'Seite',
-    'dashboard_log_message_delete' => 'Eineb Datensatz in <strong>{container}</strong> gelöscht.',
+    'dashboard_log_message_delete' => 'Ein Datensatz in <strong>{container}</strong> wurde gelöscht.',
+    'model_tag_relations_count' => 'Verwendungen',
 ];