+++ src/messages/tr/admin.php
@@ -366,4 +366,5 @@ return [
     'crud_exportdata_btn_downloadexport' => 'Dışa aktarma indir',
     'js_pagination_page' => 'Page',
     'dashboard_log_message_delete' => 'Deleted a record in <strong>{container}</strong>.',
+    'model_tag_relations_count' => 'Used',
 ];